Added RoverPivotTest testing OpMode
I'm going to work on a cleaner method for pivoting the wheel subassemblies in here. Instead of doing the movement inside a method that stops all other commands from updating, this should be able to work inside of the loop() method without needing to finish before continuing to the next loop. (Makes perfect sense I swear)